Move environment variables support into env_config.py
We will add more environment variables in the future, so better have them all in a single place.
